pyproject.toml
[project]
name = "proxyswitch"
version = "0.1.2"
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "proxy-py>=2.4.4",
]

[project.scripts]
proxyswitch = 'proxyswitch:main'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/proxyswitch"]
